update: switch to using relative directories
It's painful to rely on absolute names because we have to convert the
slashes in path names, log files become ugly, and we don't have atomic
error handling.

This new version uses relative directories instead, which means that
the directories passed in argument will have their names replicated
as-is on the other side. This means that the backup directory will
naturally contain the same directory structure as the local one, which
in the end is more convenient for restore operations as symlinks are
correctly preserved.