update: normalize the path to the modules to support slashes
Sometimes it's convenient to support slashes for backup and log modules so
that they both point to the same module but in a different sub-directory.
This used not to work only because of the symlink used as a reference, so
let's extract what remains after the first "/" in the module name as a
prefix and use it for --link-dest.