update: support listing of existing backups via -L
This one executes --list-only to retrieve the list of remote backups.
It also shows the LAST entry without seeing where it points to (but
its creation date gives a clue). Also the reported dates for each
backup directory are in fact the directory's date as saved, so they
can be confusing.