update: support purging older backups
With argument "-k" it's possible to indicate a list of periods and the
number of backups that are desired to keep over these periods. By
default after a backup, if these periods are listed, a list of the
outdated entries to be removed is listed. If "-P" is specified in
addition to these periods, all listed backups are also removed from
the backup server.

The algorithm ensures that backups elected for removal are the ones
falling out of any covered window, and do not serve to cover for a
missing backup in an adjacent window. For example if it is required
that 3 backups per week are kept and only 2 are found, an existing
backup dated 10 days will still be kept if the larger & older period
supports keeping such backups.

The periods are incremental, so if a first period indicates 7 days
and a second one 24 days, this second one will cover day 8 to day
31 and will effectively cover one month.

Failed backups are considered as a complement for successful ones,
with a lower priority. They're automatically removed once there are
enough good ones, but in the absence of good backups, they work exactly
like the good ones.

The script also supports to be called with only -k and -P to only
cause a purge to be performed. That may be used when disks are full
and shorter periods are necessary to save some space.