update: change the logic to add a link for valid backups instead
By adding a "-FAILED" link to failed backups, we were risking not to
mark as failed a backup which failed on low disk space or any I/O
error because it would not be possible to add this link.

For this reason instead now we add an "-OK" suffix to a successful
backup. This ensures that everything went well during all the process.
Suffix-less backups are now the ones considered as failed. This
provides a much more reliable solution.