update: rearrange the final backup sequence
Now the main part is cleanly split between preparation and backup. The
purpose will be to be able to go on with the old backup purge afterwards.
The error processing is now cleaner as we don't exit on error anymore,
and instead return and clean up what needs to be cleaned.